+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## 24.7.x
+* Added a new configuration option `enableTemporaryDeviceIDMode` to 'CountlyConfig' interface
+* Introduced a new `deviceID` interface for grouping device ID management related methods:
+  * `setID`
+  * `changeWithMerge`
+  * `changeWithoutMerge`
+  * `getID`
+  * `getIDType`
+  * `enableTemporaryIDMode`
+* Deprecated the following methods:
+  * `getCurrentDeviceId`
+  * `getDeviceIDType`
+  * `changeDeviceId`
+
+## 24.7.x-np
+* Added a new configuration option `enableTemporaryDeviceIDMode` to 'CountlyConfig' interface
+* Introduced a new `deviceID` interface for grouping device ID management related methods:
+  * `setID`
+  * `changeWithMerge`
+  * `changeWithoutMerge`
+  * `getID`
+  * `getIDType`
+  * `enableTemporaryIDMode`
+* Deprecated the following methods:
+  * `getCurrentDeviceId`
+  * `getDeviceIDType`
+  * `changeDeviceId`
+
 ## 24.7.0
 * Added support for the automatic sending of user properties to the server without requiring an explicit call to the `save` method to enhance data fidelity.
 * Added support for `List` values in user given segmentations for events and views.